Docker-compose利用時のエラーを解消、ディレクトリ変更、Dockerfileパス＆Docker-composeパス変更、Reactの起動を確認
