+++ client/entrypoint.sh
@@ -1,5 +1,4 @@
 #!/usr/bin/env bash
-
 npm install --flat --frozen-lockfile
 
 exec "$@" # CMDに指定されているコマンドを実行
