+++ server/entrypoint.sh
@@ -1,3 +1,4 @@
+#!/bin/bash
 go mod tidy
 
 exec "$@"
